WICKET-3386 org/apache/wicket/util/io/Streams.java copy methods return long and/or int
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1063875 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket-util/src/main/java/org/apache/wicket/util/io/Streams.java
@@ -43,7 +43,7 @@ public final class Streams
 	 * @return Number of bytes copied from one stream to the other
 	 * @throws IOException
 	 */
-	public static long copyAndClose(final InputStream in, final OutputStream out)
+	public static int copyAndClose(final InputStream in, final OutputStream out)
 		throws IOException
 	{
 		try
